reorganize all svelte content
