--- frontend/components/src/infopanel-main.js
@@ -1,9 +1,0 @@
-import './css/interface.css';
-import InfoPanel from './InfoPanel.svelte';
-
-const app = new InfoPanel({
-	target: document.getElementById("infopanel-target"),
-	props: JSON.parse(document.getElementById("infopanel-props").textContent),
-});
-
-export default app;
