+++ frontend/svelte/package.json
