+++ frontend/svelte/src/components/Footer.svelte
