+++ frontend/svelte/src/components/InfoPanel.svelte
