+++ frontend/svelte/src/components/LatestAdditions.svelte
