+++ frontend/svelte/src/components/Navbar.svelte
