+++ frontend/svelte/src/components/Participants.svelte
