+++ frontend/svelte/src/components/Place.svelte
