+++ frontend/svelte/src/components/PlaceSelect.svelte
