+++ frontend/svelte/src/components/Places.svelte
