+++ frontend/svelte/src/components/RecentActivity.svelte
