+++ frontend/svelte/src/components/TitleBar.svelte
