+++ frontend/svelte/src/css/ol-overrides.css
