+++ frontend/svelte/src/js/ol-styles-georeference.js
