+++ frontend/svelte/src/js/ol-styles.js
