+++ frontend/svelte/src/js/ol-utils-georeference.js
