+++ frontend/svelte/src/js/ol-utils.js
