+++ frontend/svelte/src/main/footer.js
@@ -1,5 +1,5 @@
-import './css/shared.css';
-import Footer from './Footer.svelte';
+import '../css/shared.css';
+import Footer from '../components/Footer.svelte';
 
 const footer = new Footer({
 	target: document.getElementById("footer-target"),
