+++ frontend/svelte/src/main/georeference.js
@@ -1,8 +1,8 @@
-import './css/interface.css';
-import './js/ol-styles.js';
-import './js/ol-utils.js';
-import Georeference from './Georeference.svelte';
-import './css/ol-overrides.css'
+import '../css/interface.css';
+import '../js/ol-styles.js';
+import '../js/ol-utils.js';
+import Georeference from '../components/Georeference.svelte';
+import '../css/ol-overrides.css'
 
 const app = new Georeference({
 	target: document.getElementById("georeference-target"),
