+++ frontend/svelte/src/main/index.js
@@ -1,4 +1,4 @@
-import Index from './Index.svelte';
+import Index from '../components/Index.svelte';
 
 const index = new Index({
 	target: document.getElementById("index-target"),
