+++ frontend/svelte/src/main/place.js
@@ -1,6 +1,6 @@
-import '../../static/css/site_base.css';
-import './css/shared.css';
-import Place from './Place.svelte';
+import '../../../static/css/site_base.css';
+import '../css/shared.css';
+import Place from '../components/Place.svelte';
 
 const place = new Place({
 	target: document.getElementById("place-target"),
