+++ frontend/svelte/src/main/resource.js
@@ -1,5 +1,5 @@
-import './css/interface.css';
-import Resource from './Resource.svelte';
+import '../css/interface.css';
+import Resource from '../components/Resource.svelte';
 
 const app = new Resource({
 	target: document.getElementById("resource-target"),
