+++ frontend/svelte/src/main/split.js
@@ -1,6 +1,6 @@
-import './css/interface.css';
-import Split from './Split.svelte';
-import './css/ol-overrides.css'
+import '../css/interface.css';
+import Split from '../components/Split.svelte';
+import '../css/ol-overrides.css'
 
 const app = new Split({
 	target: document.getElementById("split-target"),
